// src/app/shared/models/bandejaMatrizRiesgo.ts
export class BandejaMatrizRiesgo {

    idMatrizRiesgo?: number;
    idCartera?: number;
    deCartera?: string;
    idEmpresa?: number;
    nombreCortoEmpresa?: string;
    idPeriodo?: number;
    dePeriodo?: string;
    idTipoMatriz?: number;
    deTipoMatriz?: string;
    idMatrizNivel?: number;
    deMatrizNivel?: string;
    fechaCreacion?: string;
    indicadorBaja?: string;

}